+++ solid-principles/typescript/srp/src/account-service.ts
@@ -1,10 +1,9 @@
-import TransactionRepository from './transaction-repository';
+import TransactionRepository, {FormatTransactions} from './transaction-repository';
 import Clock from './clock';
 import Console from './console';
 import Transaction from './transaction';
 
 class AccountService {
-  private STATEMENT_HEADER: string = 'DATE | AMOUNT | BALANCE';
 
   private transactionRepository: TransactionRepository;
   private clock: Clock;
@@ -29,48 +28,14 @@ class AccountService {
   }
 
   public printStatement(): void {
-    this.printHeader();
-    this.printTransactions();
+    const formatTransaction = new FormatTransactions(this.transactionRepository.all());
+    const lines = formatTransaction.print();
+    lines.forEach(line => this.console.printLine(line));
   }
 
-  private printHeader() {
-    this.printLine(this.STATEMENT_HEADER);
-  }
-
-  private printTransactions() {
-    const transactions: Transaction[] = this.transactionRepository.all();
-    let balance = 0;
-
-    transactions
-      .map(transaction => {
-        balance += transaction.getAmount();
-        return this.statementLine(transaction, balance);
-      })
-      .forEach(statement => this.printLine(statement));
-  }
-
+  // ===== Private =====
   private transactionWith(amount: number): Transaction {
     return new Transaction(this.clock.today(), amount);
-  }
-
-  private statementLine(transaction: Transaction, balance: number) {
-    const formattedDate = this.formatDate(transaction.getDate());
-    const formattedAmmount = this.formatNumber(transaction.getAmount());
-    const formattedBalance = this.formatNumber(balance);
-
-    return `${formattedDate} | ${formattedAmmount} | ${formattedBalance}`;
-  }
-
-  private formatDate(date: Date): string {
-    return `${date.getDate()}/${date.getMonth() + 1}/${date.getFullYear()}`;
-  }
-
-  private formatNumber(amount: number): string {
-    return amount.toFixed(2);
-  }
-
-  private printLine(line: string) {
-    this.console.printLine(line);
   }
 }
